Caps and Crafting, Third Attempt
Reworked the trading vendor to be more simple, and allow for easier addition of value to items - currently, only the ores, the chems it used to take, and the new crafting components have value.
Added bones to the BBQ ribs recipe in place of rods.
Re-added several old foods and their recipes from before the rebase.