Food: Removes Bookcrafting from the Food.
## About The Pull Request

Basically edits some of the food code, to make it not locked behind books, so like the banana split, some of the shakes, vanilla cake, clown cake, mime tart, and more. This way it is more known for people ot actually use them instead of sititng in the code.

## Why It's Good For The Game

This is good for the game as it fixes the food and readds these per say, into the game so they are craftable by every player instead of looking for a book that does not exist.

## Changelog
:cl:
tweak: tweaks and removes some of the line of the code so that it was being used by books that you cannot obtain without admin intervision and should be able to be craftable now, without books awhile keeping them in the game for future use if needed.
/:cl: