Sprites! Sidearms and shorts! Lasers!
I made a ton of sprites. Made the Wattz rifle and pistol and both their upgraded variants more game-accurate, did the same for the 6520, made a bunch of pairs of short shorts, and made sprites for ranger versions of the sidearms to be used in a later update that makes them a separate weapon after balance team finishes the big weapon stat overhaul. The shorts are a loadout option, too.