Removes My Ranger Sidearm Sprites
Ranger sidearms are probably being removed as a concept entirely, and my project was given a no on top of that. It's really cool to know I'm technically the reason the ranger takedowns got removed and the sidearms too. At least the takedown thing was because of something someone else did to me, instead of something I was trying to improve. This entire venture was worse than a waste of time, because at least a waste of time wouldn't have ruined the thing I was trying to improve, it would've simply left it be. Everything I try to do is reversed into something awful. Being a part of this server is a lot of fun.

I refuse the use of these sprites as any sort of alts to the normal revolvers because that's so far removed from their initial concept. Don't re-add them.