My custom holster and gun description changes
One part of this is adding my character's NCRCF themed holster. The second part is something that I've hoped to get changed for a while. Some of the guns haven't been sprite-accurate in description or name and I've been trying to fix it for a while, Raptor told me he'd get it in but it's been a while so I decided to slap it into this PR while I was at it. I wrote new descriptions for the Beretta and the .45 revolver for now, and I'm willing to look into a proper Casull resprite for the wasteland .45 if asked in the future should the old description be preferable.