Pisshole's NCR Update, Part 1
Entirely reworks the NCR Medical Officer in to the OSI Officer, including a rename, a change in clothing and three loadouts. This is in part preparation for a new NCR role planned down the line. Also - fixes NCR Sergeant First Class to NCR First Sergeant, for lore reasons, and cleans up/fixes a few lines of code including ones relating to redundant NCR CO entries and missing loadout options for the NCR Ranger Sergeant.